# docs/swagger.yaml
basePath: /
definitions:
  book.CreateBookReq:
    properties:
      author:
        example: John Doe
        type: string
      title:
        example: The Great Gatsby
        type: string
    required:
    - author
    - title
    type: object
  book.UpdateBookReq:
    properties:
      author:
        example: John Doe
        type: string
      title:
        example: The Great Gatsby
        type: string
    required:
    - author
    - title
    type: object
  domain.Book:
    properties:
      author:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  restkit.Error:
    properties:
      code:
        $ref: '#/definitions/restkit.ErrorCode'
      message:
        type: string
      status:
        type: integer
    type: object
  restkit.ErrorCode:
    enum:
    - INTERNAL_ERROR
    - VALIDATION_FAILED
    - NOT_FOUND
    type: string
    x-enum-varnames:
    - ErrCodeInternalError
    - ErrCodeValidationFailed
    - ErrCodeNotFound
host: localhost:8080
info:
  contact: {}
  description: This is a sample server API Demo.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: API Demo
  version: "1.0"
paths:
  /books:
    get:
      consumes:
      - application/json
      description: Get books
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per Page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Book'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/restkit.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/restkit.Error'
      summary: Get books
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Create a book
      parameters:
      - description: Create book
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/book.CreateBookReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/restkit.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/restkit.Error'
      summary: Create a book
      tags:
      - books
  /books/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/restkit.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/restkit.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/restkit.Error'
      summary: Delete a book
      tags:
      - books
    get:
      consumes:
      - application/json
      description: Get a book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/restkit.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/restkit.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/restkit.Error'
      summary: Get a book
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Update a book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update book
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/book.UpdateBookReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/restkit.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/restkit.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/restkit.Error'
      summary: Update a book
      tags:
      - books
swagger: "2.0"